+++ ExpandWorldSize/ExpandWorldSize.cs
@@ -12,7 +12,7 @@ public class EWS : BaseUnityPlugin
 {
   public const string GUID = "expand_world_size";
   public const string NAME = "Expand World Size";
-  public const string VERSION = "1.13";
+  public const string VERSION = "1.14";
 #nullable disable
   public static ManualLogSource Log;
 #nullable enable